workspace-setup/terraform-checker/requirements.txt
# AWS SDK
boto3>=1.34.0

# Azure SDKs
azure-identity>=1.15.0
azure-mgmt-resource>=23.0.0
azure-mgmt-network>=25.0.0
azure-mgmt-storage>=21.0.0
azure-mgmt-keyvault>=10.0.0
azure-mgmt-authorization>=4.0.0
azure-mgmt-privatedns>=1.0.0
azure-mgmt-compute>=30.0.0

# GCP SDKs
google-cloud-storage>=2.14.0
google-cloud-compute>=1.14.0
google-cloud-resource-manager>=1.12.0
google-cloud-kms>=2.21.0
google-api-python-client>=2.111.0

# CLI and utilities
click>=8.1.0
rich>=13.7.0
pyyaml>=6.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

